requirements.txt
fastapi
uvicorn
pydantic
diffusers
transformers
accelerate
torch
torchvision
Pillow
numpy
requests
python-multipart
python-dotenv
google-genai